refactor auth file structure
