--- src/components/auth/index.ts
@@ -1,4 +1,0 @@
-import LoginForm from './LoginForm';
-import SignUpForm from './SignupForm';
-
-export { LoginForm, SignUpForm };
