+++ src/screens/AuthLayout.tsx
@@ -4,11 +4,11 @@ import React from 'react';
 
 /**
  * Wrapper for all authentication components
- * LoginForm, SignupForm
+ * LoginForm, SignUpForm
  *
  * @returns {JSX.Element}
  */
-const AuthScreen: React.FC = (): JSX.Element => {
+const AuthLayout: React.FC = (): JSX.Element => {
     const { session } = useSessionContext();
 
     return (
@@ -18,4 +18,4 @@ const AuthScreen: React.FC = (): JSX.Element => {
     );
 };
 
-export default AuthScreen;
+export default AuthLayout;
